requirements.txt
aiogram==2.25.1
python-dotenv==1.0.0
APScheduler==3.9.1
SQLAlchemy==1.4.41
